requirements.txt
asttokens==3.0.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==5.2.1
executing==2.2.0
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
lxml==5.4.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.4
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
playwright==1.54.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.1
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
yt-dlp==2025.3.31
